Argh! And now without a warning..
git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@957 561b6253-61b6-49b5-99ee-40376865e0bd

+++ src/items.c
@@ -304,7 +304,7 @@ item *item_new_by_level(item_t item_type, int num_level)
     /* clean results */
     if (id_min < 1) id_min = 1;
     if (id_max < 1) id_max = 1;
-    if (id_max > item_max_id(item_type)) id_max = item_max_id(item_type);
+    if (id_max > (int)item_max_id(item_type)) id_max = item_max_id(item_type);
 
     /* create the item */
     nitem = item_new(item_type, rand_m_n(id_min, id_max));
